+++ CHANGELOG.md
@@ -1,3 +1,5 @@
-## [0.0.1] - TODO: Add release date.
+## 1.0.0-nullsafety - 2021.01.24
 
-* TODO: Describe initial release.
+ - Included behavior subject and yours dependencies to local directory;
+ - Added IPersistenceSubject and your basic implementations;
+ - Added IPersistence and your basic implementations;